refactor common methods to utils
